requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
Flask-Cors==4.0.1
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
werkzeug==3.0.3
zipp==3.19.2
